Moving useUpdateCache to hooks folder as more components are using it.
